finally posting it
This is what I've done so far. I've developed all of this on Ubuntu,
but that distro makes me want to kill myself, so I've decided that
in the interest of my own safety and wellbeing, I will transfer
development to my main Arch PC. I thought I might as well commit
everything I've done to GitHub while I was at it, because this repo
has been empty for quite a while now.

Signed-off-by: RogueAI42 <declanhoare@exemail.com.au>