Fixed enough to launch
The application now launches on Arch! I've still got some work to
do to get rid of the ceaseless "from module import *" in some
files, and exception handling is also very quiet, which I'm
trying to resolve. This is one pretzel of a program but I'm going
to fix it. I removed some of the script entries from
script_data.xml because they were in the wrong format, but this is
just for testing purposes - I'll re-add them once I've got the new
backend working.